remark view as table; misc

+++ src/main/kotlin/seepick/localsportsclub/view/common/CustomDialog.kt
@@ -4,7 +4,7 @@ import androidx.compose.runtime.Composable
 
 data class CustomDialog(
     val title: String,
-    val content: @Composable () -> Unit,
+    val content: @Composable (Int) -> Unit,
     val confirmLabel: String = "Close",
     val confirmButtonTooltip: String? = null,
     val onConfirm: () -> Unit = {},
